Add Electron build configuration
