+++ index.html
@@ -33,6 +33,6 @@
     <iframe id="previewFrame" style="display:none;"></iframe>
     <div class="footer">Développé avec ❤️ par Jeff Longele — vos fichiers ne sont jamais enregistrés.</div>
   </div>
-  <script src="main.js"></script>
+  <script src="renderer.js"></script>
 </body>
 </html>
