+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "fusionx",
+  "version": "1.0.0",
+  "description": "FusionX desktop build",
+  "main": "main.js",
+  "scripts": {
+    "start": "electron .",
+    "pack": "electron-builder --dir",
+    "dist": "electron-builder"
+  },
+  "devDependencies": {
+    "electron": "^28.0.0",
+    "electron-builder": "^24.6.4"
+  },
+  "build": {
+    "appId": "com.example.fusionx",
+    "productName": "FusionX",
+    "files": [
+      "**/*",
+      "!dist/**"
+    ],
+    "directories": {
+      "output": "dist"
+    },
+    "win": {
+      "target": "nsis"
+    }
+  }
+}
