+++ package.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-mkdir -p dist && zip -r dist/FusionX.zip index.html style.css main.js fonts README.md
+mkdir -p dist && zip -r dist/FusionX.zip index.html style.css renderer.js fonts README.md
